% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{after_thing}
\alias{after_thing}
\title{Check if a node is after another node}
\usage{
after_thing(body, thing = "code_block")
}
\arguments{
\item{body}{an XML node}

\item{thing}{the name of the XML node for the node to be after,
defaults to "code_block"}
}
\value{
a single boolean value indicating if the node has a
single sibling that is a code block
}
\description{
Check if a node is after another node
}
\keyword{internal}
