% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_blocks}
\alias{are_blocks}
\title{test if the children of a given nodeset are kramdown blocks}
\usage{
are_blocks(krams)
}
\arguments{
\item{krams}{a nodeset}
}
\value{
a boolean vector equal to the length of the nodeset
}
\description{
test if the children of a given nodeset are kramdown blocks
}
\keyword{internal}
