% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{clean_native_divs}
\alias{clean_native_divs}
\title{Clean pandoc native divs and place them in their own paragraph elements}
\usage{
clean_native_divs(body)
}
\arguments{
\item{body}{an xml document}
}
\value{
an xml document
}
\description{
Sometimes pandoc native divs are bunched together, which makes it difficult
to track the pairs. This separates them into different paragraph elements so
that we can track them
}
\examples{
txt <- "::::::: challenge
## Challenge

do that challenging thing.

```{r}
cat('it might be challenging to do this')
```
::::: solution ::::
```{r}
It's not that challenging
```
::::
::: solution ::::::::
We just have to try harder and use `<div>` tags

```{r}
cat('better faster stronger with <div>')
```
<img src='https://carpentries.org/logo.svg'/>

What if we include some `:::` code in here or ::: like this

:::::
:::::

::: good

## Good divs

:::"
f <- tempfile()
writeLines(txt, f)
ex <- tinkr::to_xml(f, sourcepos = TRUE)
ex$body
predicate <- ".//d1:paragraph[d1:text[starts-with(text(), ':::')]]"
xml2::xml_text(xml2::xml_find_all(ex$body, predicate))
pegboard:::clean_native_divs(ex$body)
xml2::xml_text(xml2::xml_find_all(ex$body, predicate))
}
\seealso{
\code{\link[=get_divs]{get_divs()}} for finding labelled tags,
\code{\link[=find_between_tags]{find_between_tags()}} to extract things between the tags,
\code{\link[=label_div_tags]{label_div_tags()}} for labelling div tags,
\code{\link[=replace_with_div]{replace_with_div()}} for replacing blockquotes with div tags
\code{\link[=find_div_pairs]{find_div_pairs()}} for connecting open and closing tags
\code{\link[=clean_native_divs]{clean_native_divs()}} for cleaning cluttered pandoc div tags
}
\keyword{internal}
