% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_helpers.R
\name{issue_warning}
\alias{issue_warning}
\alias{append_labels}
\title{Issue a warning via CLI if it exists or send a message}
\usage{
issue_warning(msg, cli = FALSE, ...)

append_labels(l, i = TRUE, e = "", cli = FALSE, f = "style_inverse")
}
\arguments{
\item{msg}{the message as a glue or CLI string}

\item{cli}{if \code{TRUE}, stylizes \code{e} with \code{f}}

\item{...}{named arguments to be evaluated in the message via glue or CLI}

\item{l}{a vector/list of characters}

\item{i}{the index of elements to append}

\item{e}{the new element to append to each element}

\item{f}{a function from \pkg{cli} that will transform \code{e}}
}
\value{
nothing, invisibly; used for side-effect

, \code{l}, appended
}
\description{
Issue a warning via CLI if it exists or send a message

Append a stylized label to elements of a vector
}
\examples{
x <- letters[1:5]
x2 <- pegboard:::append_labels(x, 
  c(1, 3), 
  "appended", 
  cli = requireNamespace("cli", quietly = TRUE), 
  f = "col_cyan"
)
cat(glue::glue("[{x}]->[{x2}]"))
}
