% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_roxygen.R
\name{convert_to_roxygen}
\alias{convert_to_roxygen}
\title{Convert a given block quote to roxygen template}
\usage{
convert_to_roxygen(block, token = "#'")
}
\arguments{
\item{block}{a block quote element}

\item{token}{the token to use to indicate markdown text over code}
}
\value{
the block, converted to a code block
}
\description{
Convert a given block quote to roxygen template
}
\examples{
frg <- Lesson$new(lesson_fragment())
blo <- frg$episodes$`14-looping-data-sets.md`$get_blocks()[[2]]
convert_to_roxygen(blo)
}
\keyword{internal}
