% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevate_children.R
\name{elevate_children}
\alias{elevate_children}
\title{elevate all children of a node}
\usage{
elevate_children(parent, remove = TRUE)
}
\arguments{
\item{parent}{an xml node (notably a block quote)}

\item{remove}{a logical value. If \code{TRUE} (default), the parent node is
removed from the document.}
}
\value{
the elevated nodes, invisibly
}
\description{
elevate all children of a node
}
\examples{
scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
# get all the challenges (2 blocks)
scope$get_blocks(".challenge")
b1 <- scope$get_blocks(".challenge")[[1]]
elevate_children(b1)
# now there is only one block:
scope$get_blocks(".challenge")
}
\keyword{internal}
