% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lesson.R
\name{Episode}
\alias{Episode}
\title{Class representing XML source of a Carpentries episode}
\description{
Wrapper around an xml document to manipulate and inspect Carpentries episodes
}
\details{
This class is a fancy wrapper around the results of \code{\link[tinkr:to_xml]{tinkr::to_xml()}} and
has method specific to the Carpentries episodes.
}
\examples{

## ------------------------------------------------
## Method `Episode$challenges`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$challenges()

## ------------------------------------------------
## Method `Episode$name`
## ------------------------------------------------

scope$name()

## ------------------------------------------------
## Method `Episode$lesson`
## ------------------------------------------------

scope$lesson()

## ------------------------------------------------
## Method `Episode$new`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{[\code{character}] path to file on disk}

\item{\code{yaml}}{[\code{character}] text block at head of file}

\item{\code{body}}{[\code{xml_document}] an xml document of the episode}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-challenges}{\code{Episode$challenges()}}
\item \href{#method-name}{\code{Episode$name()}}
\item \href{#method-lesson}{\code{Episode$lesson()}}
\item \href{#method-new}{\code{Episode$new()}}
\item \href{#method-clone}{\code{Episode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-challenges"></a>}}
\if{latex}{\out{\hypertarget{method-challenges}{}}}
\subsection{Method \code{challenges()}}{
Extract all challenge blocks from the Episode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Episode$challenges()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an \code{xml_nodeset} object with each node being a \code{block_quote}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$challenges()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-name"></a>}}
\if{latex}{\out{\hypertarget{method-name}{}}}
\subsection{Method \code{name()}}{
name of the file without the path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Episode$name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{scope$name()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lesson"></a>}}
\if{latex}{\out{\hypertarget{method-lesson}{}}}
\subsection{Method \code{lesson()}}{
path to the lesson
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Episode$lesson()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{scope$lesson()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Episode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Episode$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{[\code{character}] path to a markdown episod file on disk}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Episode object with extraced XML data
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Episode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
