% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{find_between_tags}
\alias{find_between_tags}
\title{Find nodes between two nodes with a given dtag}
\usage{
find_between_tags(
  tag,
  body,
  ns = "pb",
  find = "dtag[@label='{tag}']",
  include = FALSE
)
}
\arguments{
\item{tag}{an integer representing a unique dtag attribute}

\item{body}{an xml document}

\item{ns}{the namespace from the body}

\item{find}{an xpath element to search for (without namespace tag)}

\item{include}{if \code{TRUE}, the tags themselves will be included in the output}
}
\value{
a nodeset between tags that have the dtag attribute matching \code{tag}
}
\description{
Find nodes between two nodes with a given dtag
}
\examples{
loop <- Episode$new(file.path(lesson_fragment(), "_episodes", "14-looping-data-sets.md"))
loop$body # a full document with block quotes and code blocks, etc
loop$unblock() # removing blockquotes and replacing with div tags
# find all the div tags
ns <- pegboard:::get_ns(loop$body)
tags <- xml2::xml_find_all(loop$body, ".//pb:dtag", ns)
tags <- xml2::xml_attr(tags, "label")
tags
# grab the contents of the first div tag
pegboard:::find_between_tags(tags[[1]], loop$body)
}
\seealso{
\code{\link[=get_divs]{get_divs()}} for finding labelled tags,
\code{\link[=find_between_tags]{find_between_tags()}} to extract things between the tags,
\code{\link[=label_div_tags]{label_div_tags()}} for labelling div tags,
\code{\link[=clean_div_tags]{clean_div_tags()}} for cleaning cluttered div tags,
\code{\link[=replace_with_div]{replace_with_div()}} for replacing blockquotes with div tags
\code{\link[=find_div_pairs]{find_div_pairs()}} for connecting open and closing tags
\code{\link[=clean_fenced_divs]{clean_fenced_divs()}} for cleaning cluttered pandoc div tags
}
\keyword{div}
\keyword{internal}
