% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blocks.R
\name{get_blocks}
\alias{get_blocks}
\title{Gather blocks from the XML body of a carpentries lesson}
\usage{
get_blocks(body, type = NULL, level = 1)
}
\arguments{
\item{body}{the XML body of a carpentries lesson (an xml2 object)}

\item{type}{the type of block quote in the Jekyll syntax like ".challenge",
".discussion", or ".solution"}

\item{level}{the level of the block within the document. Defaults to \code{1},
which represents all of the block_quotes are not nested within any other
block quotes. Increase the nubmer to increase the level of nesting.}
}
\value{
an xml nodeset object with each element representing a blockquote
that matched the input criteria.
}
\description{
This will search an XML document for \code{block_quotes} with the specified type
and level and extract them into a nodeset.
}
\note{
At the moment, blocks are returned at the specified level. If you
select \verb{type = ".solution", level = 1}, you will receive blocks that
\emph{contain} solution blocks even though these blocks are almost always nested
within other blocks.
}
\examples{
frg <- Lesson$new(lesson_fragment())
# Find all the blocks in the
get_blocks(frg$episodes[["17-scope.md"]]$body)
}
