% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_challenges.R
\name{get_challenges}
\alias{get_challenges}
\title{Gather challenges from the XML body of a carpentries lesson}
\usage{
get_challenges(body, type = c("block", "div", "chunk"))
}
\arguments{
\item{body}{the XML body of a carpentries lesson (an xml2 object)}

\item{type}{the type of element containing the challenges "block" is the
default and will search for all of the blockquotes with liquid/kramdown
markup, "div" will search for all div tags with class of challenge, and
"chunk" will search for all of code chunks with the engine of challenge.}
}
\value{
an xml object.
}
\description{
This will search an XML document for a challenge marker and extract all of
the block quotes that are ancestral to that marker so that we can extract the
challenge blockquotes from the carpentries lessons.
}
\examples{
loop <- Episode$new(file.path(lesson_fragment(), "_episodes", "14-looping-data-sets.md"))
get_challenges(loop$body, "block")
get_challenges(loop$unblock()$body, "div")
loop$reset()
get_challenges(loop$use_dovetail()$unblock()$body, "chunk")
}
