% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code}
\alias{get_code}
\title{Get code blocks from xml document}
\usage{
get_code(body, type = ".language-", attr = "@ktag")
}
\arguments{
\item{body}{an xml document from a jekyll site}

\item{type}{a full or partial string of a code block attribute from Jekyll
without parenthesis.}

\item{attr}{what attribute to query in search of code blocks. Default is
@ktag, which will search for "\{: \<type\>".}
}
\value{
an xml nodeset object
}
\description{
Get code blocks from xml document
}
\details{
This uses the XPath function \code{fn:starts-with()} to  search for the
code block and automatically includes the opening brace, so regular
expressions are not allowed. This is used by the \verb{$code}, \verb{$output}, and
\verb{$error} elements of the \link{Episode} class.
}
\examples{

e <- Episode$new(fs::path(lesson_fragment(), "_episodes", "17-scope.md"))

get_code(e$body)
get_code(e$body, ".output")
get_code(e$body, ".error")
}
