% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{get_div_levels}
\alias{get_div_levels}
\title{Get levels of a character vector of div tags}
\usage{
get_div_levels(nodes)
}
\arguments{
\item{nodes}{a character vector of div open and close tags}
}
\value{
an integer vector indicating the depth of the tags where 0 indicates
a closing tag
}
\description{
Get levels of a character vector of div tags
}
\examples{
nodes <- c(
"<div class='1'>", 
  "<div class='2'>" , 
  "</div>", 
  "<div class='2'>", 
    "<div class='3'>", 
    "</div>", 
  "</div>", 
"</div>")
pegboard:::get_div_levels(nodes)
}
\keyword{internal}
