% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{get_divs}
\alias{get_divs}
\title{Get paired div blocks}
\usage{
get_divs(body, type = NULL, include = FALSE)
}
\arguments{
\item{body}{an xml document}

\item{type}{the type of div to return}
}
\value{
a list of nodesets
}
\description{
Get paired div blocks
}
\details{
The strategy behind this is to :
\enumerate{
\item find all \code{dtag} elements, which will necessarily be <html_block> elements
containing div tags
\item grab the first tag of each pair
\item filter on div tag class (type)
\item grab all elements between the tags
}
}
\examples{
loop <- Episode$new(file.path(lesson_fragment(), "_episodes", "14-looping-data-sets.md"))
loop$body # a full document with block quotes and code blocks, etc
loop$unblock() # removing blockquotes and replacing with div tags
pegboard:::get_divs(loop$body, 'challenge') # all challenge blocks
pegboard:::get_divs(loop$body, 'solution') # all solution blocks
}
\seealso{
\code{\link[=get_divs]{get_divs()}} for finding labelled tags,
\code{\link[=find_between_tags]{find_between_tags()}} to extract things between the tags,
\code{\link[=label_div_tags]{label_div_tags()}} for labelling div tags,
\code{\link[=clean_div_tags]{clean_div_tags()}} for cleaning cluttered div tags,
\code{\link[=replace_with_div]{replace_with_div()}} for replacing blockquotes with div tags
\code{\link[=find_div_pairs]{find_div_pairs()}} for connecting open and closing tags
\code{\link[=clean_fenced_divs]{clean_fenced_divs()}} for cleaning cluttered pandoc div tags
}
\keyword{div}
\keyword{internal}
