% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_images}
\alias{get_images}
\title{Get images from an Episode/yarn object}
\usage{
get_images(yrn, process = TRUE)
}
\arguments{
\item{yrn}{an Episode/yarn object}

\item{process}{if \code{TRUE} (default), images will be processed via
\code{\link[=process_images]{process_images()}} to add the alt attribute and extract images from HTML
blocks. \code{FALSE} will present the nodes as found by XPath search.}
}
\value{
an xml_nodelist
}
\description{
Get images from an Episode/yarn object
}
\keyword{internal}
