% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lesson.R
\name{get_lesson}
\alias{get_lesson}
\title{Get a carpentries lesson in XML format}
\usage{
get_lesson(lesson = NULL, path = tempdir(), overwrite = FALSE, ...)
}
\arguments{
\item{lesson}{a github user/repo pattern to point to the lesson}

\item{path}{a directory to write the lesson to}

\item{overwrite}{if the \code{path} exists, setting this to \code{TRUE} will overwrite
the path, otherwise, the contents of the path will be returned if it is a
lesson repository.}

\item{...}{arguments passed on to \link[=Episode]{Episode$new()}.}
}
\value{
a list of xml objects, one element per episode.
}
\description{
Download and extract a carpentries lesson in XML format. This uses \code{\link[gert:git_fetch]{gert::git_clone()}}
to download a carpentries lesson to your computer (defaults to the temporary
directory and extracts the lesson in \verb{_episodes/} using \code{\link[tinkr:to_xml]{tinkr::to_xml()}}
}
\examples{

if (interactive()) {
  png <- get_lesson("swcarpentry/python-novice-gapminder")
  str(png, max.level = 1)
}
}
