% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_block.R
\name{get_list_block}
\alias{get_list_block}
\title{Extract list elements from a block}
\usage{
get_list_block(self, type = "questions", in_yaml = TRUE)
}
\arguments{
\item{self}{an Episode object}

\item{type}{the type of block/div to extract the list from}

\item{in_yaml}{indicator if the elements are in the YAML header (\code{TRUE},
default for styles version 9 lessons) or if they are in the body of the
lesson (\code{FALSE}, for sandpaper lessons).}
}
\value{
a character vector
}
\description{
Extract list elements from a block
}
\keyword{internal}
