% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solutions.R
\name{get_solutions}
\alias{get_solutions}
\title{Gather solutions from the XML body of a carpentries lesson}
\usage{
get_solutions(body, parent = ".challenge")
}
\arguments{
\item{body}{the XML body of a carpentries lesson (an xml2 object)}

\item{parent}{the outer block containing the solution. Default is a challenge
block, but it could also be a discussion block.}
}
\value{
an xml object.
}
\description{
This will search an XML document for a solution marker and extract all of
the block quotes that are ancestral to that marker so that we can extract the
solution blockquotes from the carpentries lessons.
}
\examples{
frg <- Lesson$new(lesson_fragment())
get_solutions(frg$episodes[["17-scope.md"]]$body)
}
