% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading_tree.R
\name{heading_tree}
\alias{heading_tree}
\title{This constructs a data frame of headings for displaying to the user
via the CLI package.}
\usage{
heading_tree(headings)
}
\arguments{
\item{headings}{an object of class ``}
}
\description{
This constructs a data frame of headings for displaying to the user
via the CLI package.
}
