% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-headings.R
\name{get_headings}
\alias{get_headings}
\alias{show_heading_tree}
\title{Get all headings in the XML document}
\usage{
get_headings(body)

show_heading_tree(tree)
}
\arguments{
\item{body}{an XML document}

\item{tree}{a data frame produced via \code{\link[=validate_headings]{validate_headings()}}}
}
\value{
an object of class \code{xml_nodeset} with all the headings in the
document.
}
\description{
Get all headings in the XML document
}
