% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_elements.R
\name{isolate_elements}
\alias{isolate_elements}
\title{Isolate elements in an XML document by source position}
\usage{
isolate_elements(body, ...)
}
\arguments{
\item{body}{an XML document}

\item{...}{objects of class \code{xml_node} or \code{xml_nodeset} to be retained}
}
\value{
This works by side-effect, but it returns the body, invisibly.
}
\description{
Isolate elements in an XML document by source position
}
