% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_kram_blocks.R
\name{isolate_kram_blocks}
\alias{isolate_kram_blocks}
\title{Isolate all blocks within a document that have a kramdown tag}
\usage{
isolate_kram_blocks(body, predicate = "")
}
\arguments{
\item{body}{an xml document}

\item{predicate}{an XPath conditional statement in square brackets or an
empty string. This can be used to filter on such attributes such as line
number.}
}
\value{
the modified body, invisibly
}
\description{
Isolate all blocks within a document that have a kramdown tag
}
\keyword{internal}
