% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kramdown_tags.R
\name{kramdown_tags}
\alias{kramdown_tags}
\title{Find all kramdown tags within the document}
\usage{
kramdown_tags(body)
}
\arguments{
\item{body}{an XML document}
}
\value{
an XML nodeset with all the kramdown tags
}
\description{
Find all kramdown tags within the document
}
\keyword{internal}
