% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{label_div_tags}
\alias{label_div_tags}
\title{Add labels to div tags in the form of a "dtag" attribute}
\usage{
label_div_tags(body)
}
\arguments{
\item{body}{an xml document}
}
\value{
the document, invisibly
}
\description{
Add labels to div tags in the form of a "dtag" attribute
}
\keyword{internal}
