% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{label_pairs}
\alias{label_pairs}
\title{Label pairs of parentheses.}
\usage{
label_pairs(pairs, n_tags, reverse = FALSE)
}
\arguments{
\item{pairs}{a vector of parentheses.}

\item{n_tags}{the number of closing tags}

\item{reverse}{if \code{TRUE}, the search from the end of the stack}
}
\value{
a vector of integers indicating the pair of parentheses.
}
\description{
This function is the labeller for \code{\link[=find_div_pairs]{find_div_pairs()}}
}
\examples{

x <- c("(", "(", ")", ")")
pegboard:::label_pairs(x, 2)
x <- c("(", "(", ")", "(", "(", ")", ")", ")")
pegboard:::label_pairs(x, 4)
}
\seealso{
Other div: 
\code{\link{clean_div_tags}()},
\code{\link{find_between_tags}()},
\code{\link{find_div_pairs}()},
\code{\link{get_divs}()},
\code{\link{label_div_tags}()},
\code{\link{make_div}()},
\code{\link{make_div_pairs}()},
\code{\link{replace_with_div}()}
}
\concept{div}
\keyword{internal}
