% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson_fragment.R
\name{lesson_fragment}
\alias{lesson_fragment}
\title{An example lesson fragment}
\usage{
lesson_fragment()
}
\value{
a path to a lesson fragment that contains one \verb{_episodes} directory
with three files: "10-lunch.md", "14-looping-data-sets.md", and
"17-scope.md"
}
\description{
This example was taken from the python novice gapminder lesson
}
\examples{

lesson_fragment()
}
