% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heading_table.R
\name{make_heading_table}
\alias{make_heading_table}
\title{Create a table for headings}
\usage{
make_heading_table(headings, offset = 5L)
}
\arguments{
\item{headings}{an \code{xml_nodeset} object with text and "level" attributes}

\item{offest}{the offset for the yaml header (artifact of {tinkr}), defaults
to 5L, which is for the two fences plus title, teaching, and exercises.}
}
\value{
a data frame with three columns:
\itemize{
\item heading the text of the heading
\item level the heading level
\item pos the position of the heading in the document
}
}
\description{
Create a table for headings
}
\examples{
path <- file.path(lesson_fragment(), "_episodes", "14-looping-data-sets.md")
loop <- Episode$new(path)
pegboard:::make_heading_table(loop$headings, offset = length(loop$yaml))
}
\keyword{internal}
