% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{make_pairs}
\alias{make_pairs}
\title{Make paired labels for opening and closing div tags}
\usage{
make_pairs(nodes)
}
\arguments{
\item{nodes}{a character vector of div open and close tags}
}
\value{
an integer vector with pairs of labels for each opening and closing
tag. Note that the labels are produced by doing a cumulative sum of the
node depths.
}
\description{
Make paired labels for opening and closing div tags
}
\examples{
nodes <- c(
"<div class='1'>", 
  "<div class='2'>" , 
  "</div>", 
  "<div class='2'>", 
    "<div class='3'>", 
    "</div>", 
  "</div>", 
"</div>")
pegboard:::make_pairs(nodes)
}
\seealso{
\code{\link[=get_divs]{get_divs()}} for finding labelled tags,
\code{\link[=find_between_tags]{find_between_tags()}} to extract things between the tags,
\code{\link[=label_div_tags]{label_div_tags()}} for labelling div tags,
\code{\link[=clean_div_tags]{clean_div_tags()}} for cleaning cluttered div tags,
\code{\link[=replace_with_div]{replace_with_div()}} for replacing blockquotes with div tags
\code{\link[=make_pairs]{make_pairs()}} for connecting open and closing tags
\code{\link[=get_div_levels]{get_div_levels()}} for documenting the level of div tags
}
\keyword{internal}
