% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pandoc_alt.R
\name{make_pandoc_alt}
\alias{make_pandoc_alt}
\title{Add alt text to images when transforming from jekyll to sandpaper}
\usage{
make_pandoc_alt(images)
}
\arguments{
\item{images}{a xml_nodeset of image nodes}
}
\value{
the images, invisibly with a new alt attribute and text removed
}
\description{
Add alt text to images when transforming from jekyll to sandpaper
}
