% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_lesson.R
\name{process_lesson}
\alias{process_lesson}
\title{Process downloaded lessons}
\usage{
process_lesson(lesson, as_list = FALSE)
}
\arguments{
\item{lesson}{a list of XML object representing lessons}

\item{as_list}{[boolean] if \code{TRUE}, the result will be converted to
a list via \code{\link[xml2:as_list]{xml2::as_list()}}. A \code{FALSE} (default) value will keep it as an
xml object.}
}
\value{
a list of XML block quotes for each episode
}
\description{
Process downloaded lessons
}
