% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_markdown_files}
\alias{read_markdown_files}
\title{Process Markdown files in a directory into Episode objects}
\usage{
read_markdown_files(src, cfg = list(), sandpaper = TRUE, ...)
}
\arguments{
\item{src}{[character] the path to a folder containing markdown episodes}

\item{cfg}{[list] a parsed config file that can be used to specify the
order of the files with a key that matches the folder name.}

\item{sandpaper}{[logical] if \code{TRUE}, the episodes are expected to be
processed with {sandpaper} and will have the \verb{$confirm_sandpaper()} method
triggered.}

\item{...}{methods passed to the \link{Episode} initializer}
}
\value{
a list of \link{Episode} objects
}
\description{
Process Markdown files in a directory into Episode objects
}
\keyword{internal}
