% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sandpaper_lesson.R
\name{read_sandpaper_lesson}
\alias{read_sandpaper_lesson}
\title{Read in all markdown files associated with a sandpaper lesson}
\usage{
read_sandpaper_lesson(path, ...)
}
\arguments{
\item{path}{the path to the lesson}

\item{...}{arguments to pass to \code{\link[=read_markdown_files]{read_markdown_files()}}}
}
\value{
a list of two elements:
\itemize{
\item episodes contains all the \link{Episode} objects representing markdown files
in the \code{episodes} folder
\item extra contains all the \link{Episode} objects representing markdown files in
all other folders (including top-level)
}
}
\description{
Read in all markdown files associated with a sandpaper lesson
}
\keyword{internal}
