% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{replace_with_div}
\alias{replace_with_div}
\title{Replace a blockquote with a div tag}
\usage{
replace_with_div(block)
}
\arguments{
\item{block}{a blockquote element}
}
\value{
the children of the element, invisibly
}
\description{
This modifies a document to replace a blockquote element with a div element
}
\examples{
frg <- Lesson$new(lesson_fragment())
lop <- frg$episodes$`14-looping-data-sets.md`
xml2::xml_find_all(lop$body, ".//d1:html_block")
lop$get_blocks(level = 1)
lop$get_blocks(level = 2)
purrr::walk(lop$get_blocks(level = 2), pegboard:::replace_with_div)
purrr::walk(lop$get_blocks(level = 1), pegboard:::replace_with_div)
lop$get_blocks()
xml2::xml_find_all(lop$body, ".//d1:html_block")
# add tags
pegboard:::label_div_tags(lop$body)
xml2::xml_find_all(lop$body, ".//d1:html_block")
}
\seealso{
\code{\link[=get_divs]{get_divs()}} for finding labelled tags,
\code{\link[=find_between_tags]{find_between_tags()}} to extract things between the tags,
\code{\link[=label_div_tags]{label_div_tags()}} for labelling div tags,
\code{\link[=clean_div_tags]{clean_div_tags()}} for cleaning cluttered div tags,
\code{\link[=replace_with_div]{replace_with_div()}} for replacing blockquotes with div tags
\code{\link[=make_pairs]{make_pairs()}} for connecting open and closing tags
\code{\link[=get_div_levels]{get_div_levels()}} for documenting the level of div tags
}
\keyword{div}
\keyword{internal}
