% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{set_alt_attr}
\alias{set_alt_attr}
\title{Set the alt text for a nodeset of images}
\usage{
set_alt_attr(images, xpath, ns)
}
\arguments{
\item{images}{a nodeset of images}

\item{xpath}{an XPath expression that finds the first curly brace immediately
after a node.}

\item{ns}{the namespace of the XML}
}
\value{
the nodeset, invisibly.
}
\description{
This finds the attribute curly braces after an image declaration, extracts
the alt text, and adds it as an attribute to the image, which is useful in
parsing the XML, and will not affect rendering.
}
\note{
this function assumes that the images entering have a curly brace
following.
}
