% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_dovetail.R
\name{to_dovetail}
\alias{to_dovetail}
\title{Convert a given block quote to dovetail template}
\usage{
to_dovetail(block, token = "#'")
}
\arguments{
\item{block}{a block quote element}

\item{token}{the token to use to indicate markdown text over code}
}
\value{
the block, converted to a code block
}
\description{
The {dovetail} package allows people to write block quotes as code blocks
formatted in {roxygen2} syntax. This internal function takes a block quote
element, elevates all the child elements, converts the block quote to a code
block and uses a custom xslt stylesheet to render the children of the block
to roxygen-formatted text and code.
}
\examples{
\dontrun{
frg <- Lesson$new(lesson_fragment())
blo <- frg$episodes$`14-looping-data-sets.md`$get_blocks()[[2]]
to_dovetail(blo)
cat(xml2::xml_text(blo))
}
}
\keyword{internal}
