% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_elements.R
\name{trim_fence}
\alias{trim_fence}
\title{Trim div fences from output}
\usage{
trim_fence(nodes)
}
\arguments{
\item{nodes}{an xml_nodeset whose first and last node are div fences}
}
\value{
the nodeset without div fences
}
\description{
Trim div fences from output
}
