% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{add_pegboard_nodes}
\alias{add_pegboard_nodes}
\title{Add a pegboard node before or after a node}
\usage{
add_pegboard_nodes(node, df)
}
\arguments{
\item{node}{a single node}

\item{df}{a data frame generated from \code{\link[=make_div_pairs]{make_div_pairs()}}}
}
\value{
NULL, invisibly
}
\description{
These nodes have the namespace of "http://carpentries.org/pegboard/"
}
\keyword{internal}
