% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Episode.R
\name{Episode}
\alias{Episode}
\title{Class representing XML source of a Carpentries episode}
\description{
Wrapper around an xml document to manipulate and inspect Carpentries episodes
}
\details{
This class is a fancy wrapper around the results of \code{\link[tinkr:to_xml]{tinkr::to_xml()}} and
has method specific to the Carpentries episodes.
}
\examples{

## ------------------------------------------------
## Method `Episode$new`
## ------------------------------------------------

scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$name
scope$lesson
scope$challenges
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{[\code{character}] path to file on disk}

\item{\code{yaml}}{[\code{character}] text block at head of file}

\item{\code{body}}{[\code{xml_document}] an xml document of the episode}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{challenges}}{[\code{xml_nodeset}] all the challenges blocks from the episode}

\item{\code{name}}{[\code{character}] the name of the source file without the path}

\item{\code{lesson}}{[\code{character}] the path to the lesson where the episode is from}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Episode$new()}}
\item \href{#method-clone}{\code{Episode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Episode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Episode$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{[\code{character}] path to a markdown episod file on disk}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Episode object with extraced XML data
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{scope <- Episode$new(file.path(lesson_fragment(), "_episodes", "17-scope.md"))
scope$name
scope$lesson
scope$challenges
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Episode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
