% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{find_between_tags}
\alias{find_between_tags}
\title{Find nodes between two nodes with a given dtag}
\usage{
find_between_tags(tag, body, ns)
}
\arguments{
\item{tag}{an integer representing a unique dtag attribute}

\item{body}{an xml document}

\item{ns}{the namespace from the body}
}
\value{
a nodeset between tags that have the dtag attribute matching \code{tag}
}
\description{
Find nodes between two nodes with a given dtag
}
\examples{
loop <- Episode$new(file.path(lesson_fragment(), "_episodes", "14-looping-data-sets.md"))
loop$body # a full document with block quotes and code blocks, etc
loop$unblock() # removing blockquotes and replacing with div tags
# find all the div tags
tags <- xml2::xml_text(xml2::xml_find_all(loop$body, ".//@dtag"))
tags
# grab the contents of the first div tag
pegboard:::find_between_tags(tags, loop$body, pegboard:::NS(loop$body))
}
\keyword{div}
\keyword{internal}
