% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{find_div_pairs}
\alias{find_div_pairs}
\title{Make paired labels for opening and closing div tags}
\usage{
find_div_pairs(divs, close = div_close_regex())
}
\arguments{
\item{close}{the regex for a valid closing tag}

\item{nodes}{a character vector of div open and close tags}
}
\value{
an integer vector with pairs of labels for each opening and closing
tag. Note that the labels are produced by doing a cumulative sum of the
node depths.
}
\description{
Make paired labels for opening and closing div tags
}
\examples{
nodes <- c(
"<div class='1'>", 
  "<div class='2'>" , 
  "</div>", 
  "<div class='2'>", 
    "<div class='3'>", 
    "</div>", 
  "</div>", 
"</div>")
pegboard:::find_div_pairs(nodes)
}
\seealso{
Other div: 
\code{\link{clean_div_tags}()},
\code{\link{find_between_tags}()},
\code{\link{get_divs}()},
\code{\link{label_div_tags}()},
\code{\link{label_pairs}()},
\code{\link{make_div}()},
\code{\link{make_div_pairs}()},
\code{\link{replace_with_div}()}
}
\concept{div}
\keyword{internal}
