% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_sandpaper_links.R
\name{fix_sandpaper_links}
\alias{fix_sandpaper_links}
\title{Fix relative and jekyll links to be compatible with {sandpaper}}
\usage{
fix_sandpaper_links(body, yml = list())
}
\arguments{
\item{body}{an XML document}

\item{yml}{the list of key/value pairs derived from the \verb{_config.yml} file}
}
\value{
the body, invisibly
}
\description{
This function will perform the transformation on three node types:
}
\details{
\itemize{
\item image
\item link
\item html_node
}

The transformation will be to remove relative paths ("../") and replace
Jekyll templating (e.g. "{{ page.root }}" and "{{ site.swc_pages }}" with
either nothing or the link to software carpentry, respectively.
}
\note{
This is absolutely NOT comprehensive and some links will fail to be
converted. If this happens, please report an issue:
\url{https://github.com/carpentries/pegboard/issues/new/}
}
\examples{

loop <- fs::path(lesson_fragment(), "_episodes", "14-looping-data-sets.md")
e <- Episode$new(loop)
b <- e$body
pegboard:::make_link_table(e)$orig
suppressWarnings(yml <- yaml::read_yaml(file.path(e$lesson, "_config.yml")))
fix_sandpaper_links(b, yml)
pegboard:::make_link_table(e)$orig
}
