% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sandpaper_links.R
\name{fix_sandpaper_links}
\alias{fix_sandpaper_links}
\title{Fix relative and jekyll links to be compatible with {sandpaper}}
\usage{
fix_sandpaper_links(body)
}
\arguments{
\item{body}{an XML node}
}
\value{
the body, invisibly
}
\description{
This function will perform the transformation on three node types:
}
\details{
\itemize{
\item image
\item link
\item html_node
}

The transformation will be to remove relative paths ("../") and replace
Jekyll templating (e.g. "{{ page.root }}" and "{{ site.swc_pages }}" with
either nothing or the link to software carpentry, respectively.
}
\examples{

loop <- fs::path(lesson_fragment(), "_episodes", "14-looping-data-sets.md")
e <- Episode$new(loop)
e$body \%>\% xml2::xml_find_all(".//d1:image")
e$body \%>\% xml2::xml_find_all(".//d1:html_block")
e$body \%>\% xml2::xml_find_all(".//d1:link[contains(@destination, '{{')]")
fix_sandpaper_links(e$body)
e$body \%>\% xml2::xml_find_all(".//d1:image")
e$body \%>\% xml2::xml_find_all(".//d1:html_block")
e$body \%>\% xml2::xml_find_all(".//d1:link[contains(@destination, '{{')]")
}
