% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_challenges.R
\name{get_challenges}
\alias{get_challenges}
\title{Gather challenges from the XML body of a carpentries lesson}
\usage{
get_challenges(body, as_list = FALSE)
}
\arguments{
\item{body}{the XML body of a carpentries lesson (an xml2 object)}

\item{as_list}{[boolean] if \code{TRUE}, the result will be converted to
a list via \code{\link[xml2:as_list]{xml2::as_list()}}. A \code{FALSE} (default) value will keep it as an
xml object.}
}
\value{
a list or xml object.
}
\description{
This will search an XML document for a challenge marker and extract all of
the block quotes that are ancestral to that marker so that we can extract the
challenge blockquotes from the carpentries lessons.
}
