% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_challenges.R
\name{get_challenges}
\alias{get_challenges}
\title{Gather challenges from the XML body of a carpentries lesson}
\usage{
get_challenges(body)
}
\arguments{
\item{body}{the XML body of a carpentries lesson (an xml2 object)}
}
\value{
an xml object.
}
\description{
This will search an XML document for a challenge marker and extract all of
the block quotes that are ancestral to that marker so that we can extract the
challenge blockquotes from the carpentries lessons.
}
\examples{
frg <- Lesson$new(lesson_fragment())
get_challenges(frg$episodes[["17-scope.md"]]$body)
}
