% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{get_div_levels}
\alias{get_div_levels}
\title{Get levels of a character vector of div tags}
\usage{
get_div_levels(nodes)
}
\arguments{
\item{nodes}{a character vector of div open and close tags}
}
\value{
an integer vector indicating the depth of the tags where 0 indicates
a closing tag
}
\description{
Get levels of a character vector of div tags
}
\examples{
nodes <- c(
"<div class='1'>", 
  "<div class='2'>" , 
  "</div>", 
  "<div class='2'>", 
    "<div class='3'>", 
    "</div>", 
  "</div>", 
"</div>")
pegboard:::get_div_levels(nodes)
}
\seealso{
\code{\link[=get_divs]{get_divs()}} for finding labelled tags,
\code{\link[=find_between_tags]{find_between_tags()}} to extract things between the tags,
\code{\link[=label_div_tags]{label_div_tags()}} for labelling div tags,
\code{\link[=clean_div_tags]{clean_div_tags()}} for cleaning cluttered div tags,
\code{\link[=replace_with_div]{replace_with_div()}} for replacing blockquotes with div tags
\code{\link[=make_pairs]{make_pairs()}} for connecting open and closing tags
}
\keyword{internal}
