% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{get_divs}
\alias{get_divs}
\title{Get paired div blocks}
\usage{
get_divs(body)
}
\arguments{
\item{body}{an xml document}
}
\value{
a list of nodesets
}
\description{
Get paired div blocks
}
\keyword{internal}
