% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{label_div_tags}
\alias{label_div_tags}
\title{Add labels to div tags in the form of a "dtag" attribute}
\usage{
label_div_tags(body)
}
\arguments{
\item{body}{an xml document}
}
\value{
the document, invisibly
}
\description{
Add labels to div tags in the form of a "dtag" attribute
}
\seealso{
\code{\link[=get_divs]{get_divs()}} for finding labelled tags,
\code{\link[=find_between_tags]{find_between_tags()}} to extract things between the tags,
\code{\link[=label_div_tags]{label_div_tags()}} for labelling div tags,
\code{\link[=clean_div_tags]{clean_div_tags()}} for cleaning cluttered div tags,
\code{\link[=replace_with_div]{replace_with_div()}} for replacing blockquotes with div tags
\code{\link[=find_div_pairs]{find_div_pairs()}} for connecting open and closing tags
\code{\link[=clean_fenced_divs]{clean_fenced_divs()}} for cleaning cluttered pandoc div tags
}
\keyword{internal}
