% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lesson.R
\name{Lesson}
\alias{Lesson}
\title{Class to contain a single Lesson by the Carpentries}
\description{
This is a wrapper for several \link{Episode} class objects.
}
\details{
Lessons are made of up several episodes within the \verb{_episodes/} directory of
a lesson. This class keeps track of several episodes and allows us to switch
between RMarkdown and markdown episodes
}
\examples{

## ------------------------------------------------
## Method `Lesson$thin`
## ------------------------------------------------

frg <- Lesson$new(lesson_fragment())
frg$thin()

## ------------------------------------------------
## Method `Lesson$new`
## ------------------------------------------------

frg <- Lesson$new(lesson_fragment())
frg$path
frg$episodes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{[\code{character}] path to Lesson directory}

\item{\code{episodes}}{[\code{list}] list of \link{Episode} class objects.}

\item{\code{rmd}}{[\code{logical}] when \code{TRUE}, the episodes represent RMarkdown
files, default is \code{FALSE} for markdown files.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{files}}{the source files for each episode}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-challenges}{\code{Lesson$challenges()}}
\item \href{#method-thin}{\code{Lesson$thin()}}
\item \href{#method-new}{\code{Lesson$new()}}
\item \href{#method-clone}{\code{Lesson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-challenges"></a>}}
\if{latex}{\out{\hypertarget{method-challenges}{}}}
\subsection{Method \code{challenges()}}{
Gather all of the challenges from the lesson in a list of xml_nodeset objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lesson$challenges(path = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{[\code{logical}] if \code{TRUE}, the names of each element
will be equivalent to the path. The default is \code{FALSE}, which gives the
name of each episode.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-thin"></a>}}
\if{latex}{\out{\hypertarget{method-thin}{}}}
\subsection{Method \code{thin()}}{
Remove episodes that have no challenges
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lesson$thin(verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{[\code{logical}] if \code{TRUE} (default), the names of each
episode removed is reported. Set to \code{FALSE} to remove this behavior.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the Lesson object, invisibly
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{frg <- Lesson$new(lesson_fragment())
frg$thin()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create a new Lesson object from a directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lesson$new(path = NULL, rmd = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{[\code{character}] path to a lesson directory. This must have a
folder called \verb{_episodes} within that contains markdown episodes}

\item{\code{rmd}}{[\code{logical}] when \code{TRUE}, the imported files will be the
source RMarkdown files. Defaults to \code{FALSE}, which reads the rendered
markdown files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new Lesson object that contains a list of \link{Episode} objects in
\verb{$episodes}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{frg <- Lesson$new(lesson_fragment())
frg$path
frg$episodes
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lesson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
