% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{make_div}
\alias{make_div}
\title{Create an xml document that contains two html_block elements
that contain div tags.}
\usage{
make_div(what)
}
\arguments{
\item{what}{the class of block}
}
\value{
an xml document with commonmark namespace
}
\description{
Create an xml document that contains two html_block elements
that contain div tags.
}
\examples{
cha <- pegboard:::make_div("challenge")
cha
cat(pegboard:::xml_to_md(cha))
}
\seealso{
\code{\link[=get_divs]{get_divs()}} for finding labelled tags,
\code{\link[=find_between_tags]{find_between_tags()}} to extract things between the tags,
\code{\link[=label_div_tags]{label_div_tags()}} for labelling div tags,
\code{\link[=clean_div_tags]{clean_div_tags()}} for cleaning cluttered div tags,
\code{\link[=replace_with_div]{replace_with_div()}} for replacing blockquotes with div tags
\code{\link[=find_div_pairs]{find_div_pairs()}} for connecting open and closing tags
\code{\link[=clean_fenced_divs]{clean_fenced_divs()}} for cleaning cluttered pandoc div tags
}
\keyword{internal}
