% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{make_div}
\alias{make_div}
\title{Create an xml document that contains two html_block elements
that contain div tags.}
\usage{
make_div(what)
}
\arguments{
\item{what}{the class of block}
}
\value{
an xml document with commonmark namespace
}
\description{
Create an xml document that contains two html_block elements
that contain div tags.
}
\examples{
cha <- pegboard:::make_div("challenge")
cha
cat(pegboard:::xml_to_md(cha))
}
\keyword{internal}
