% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_div_table.R
\name{make_div_table}
\alias{make_div_table}
\title{Create a table of divs in an episode}
\usage{
make_div_table(yrn)
}
\arguments{
\item{yrn}{a \link[tinkr:yarn]{tinkr::yarn} or \link{Episode} object.}
}
\value{
a data frame with the following columns:
\itemize{
\item path: path to the file, relative to the lesson
\item div: the type of div
\item pb_label: the label of the div
\item line: the line number of the div label
}
}
\description{
Create a table of divs in an episode
}
