% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{make_div_table}
\alias{make_div_table}
\title{Create a data frame describing the divs associated with nodes.}
\usage{
make_div_table(nodes, path = NULL, yaml = NULL)
}
\arguments{
\item{nodes}{a nodelist containing native div and fenced div tags in
\code{html_block} or \code{paragraphs}.}
}
\value{
a list of data frames for each node with the following columns:
\itemize{
\item node: numeric index of the node
\item div: the text of the individual div element, stripped of context
\item label: label of the div pair (div-label-class)
\item pos: position the label will be relative to its associated node
}
}
\description{
Native and fenced divs may have several tags grouped in a single element.
In order to mark the pairs, we need to account for what tags exist in the
nodes. This function creates that
}
\examples{
txt <- "# Example with a mix of div tags

> PLEASE NEVER DO THE LESSONS THIS WAY
>
> I AM LITERALLY JUST TESTING A TERRIBLE EXAMPLE.

--------------------------------------------------------------------------------

<div class='challenge'>
## Challenge

do that challenging thing.

```{r}
cat('it might be challenging to do this')
```
:::: solution
```{r}
It's not that challenging
```
:::
<div class='solution'>
We just have to try harder and use `<div>` tags

:::::: callout
```{r}
cat('better faster stronger with <div>')
```
::::
:::::: discussion
<img src='https://carpentries.org/logo.svg'/>
:::::
</div>
</div>

<div class='good'>

## Good divs

</div>
" 
tmp <- tempfile()
writeLines(txt, tmp)
ex <- tinkr::to_xml(tmp)
pegboard:::clean_div_tags(ex$body)
nodes <- pegboard:::find_div_tags(ex$body)
divs  <- pegboard:::make_div_table(nodes)
do.call("rbind", divs)
}
\seealso{
\code{\link[=find_div_tags]{find_div_tags()}}, \code{\link[=find_div_pairs]{find_div_pairs()}}
}
\keyword{internal}
