% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{make_pairs}
\alias{make_pairs}
\title{Make paired labels for opening and closing div tags}
\usage{
make_pairs(nodes)
}
\arguments{
\item{nodes}{a character vector of div open and close tags}
}
\value{
an integer vector with pairs of labels for each opening and closing
tag. Note that the labels are produced by doing a cumulative sum of the
node depths.
}
\description{
Make paired labels for opening and closing div tags
}
\examples{
nodes <- c(
"<div class='1'>", 
  "<div class='2'>" , 
  "</div>", 
  "<div class='2'>", 
    "<div class='3'>", 
    "</div>", 
  "</div>", 
"</div>")
pegboard:::make_pairs(nodes)
}
\keyword{internal}
