% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegboard-package.R
\docType{package}
\name{pegboard-package}
\alias{pegboard}
\alias{pegboard-package}
\title{pegboard: Explore and Manipulate Markdown Curricula from the Carpentries}
\description{
The Carpentries (<https://carpentries.org>) curricula is made of of
    lessons that are hosted as websites. Each lesson represents between a half
    day to two days of instruction and contains several episodes, which are
    written as 'kramdown'-flavored 'markdown' documents and converted to HTML
    using the 'Jekyll' static website generator. This package reads in these 
    markdown documents to 'XML' and stores them in R6 classes for convenient
    exploration and manipulation of sections within episodes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zkamvar/pegboard}
  \item Report bugs at \url{https://github.com/zkamvar/pegboard/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

}
\keyword{internal}
