% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{replace_with_div}
\alias{replace_with_div}
\title{Replace a blockquote with a div tag}
\usage{
replace_with_div(block)
}
\arguments{
\item{block}{a blockquote element}
}
\value{
the children of the element, invisibly
}
\description{
This modifies a document to replace a blockquote element with a div element
}
\examples{
frg <- Lesson$new(lesson_fragment())
lop <- frg$episodes$`14-looping-data-sets.md`
xml2::xml_find_all(lop$body, ".//d1:html_block")
lop$get_blocks(level = 1)
lop$get_blocks(level = 2)
purrr::walk(lop$get_blocks(level = 2), pegboard:::replace_with_div)
purrr::walk(lop$get_blocks(level = 1), pegboard:::replace_with_div)
lop$get_blocks()
# add tags
pegboard:::label_div_tags(lop$body)
lop$get_divs()
xml2::xml_text(lop$get_divs()[[1]])
}
\seealso{
Other div: 
\code{\link{clean_div_tags}()},
\code{\link{find_between_tags}()},
\code{\link{find_div_pairs}()},
\code{\link{get_divs}()},
\code{\link{label_div_tags}()},
\code{\link{label_pairs}()},
\code{\link{make_div_pairs}()},
\code{\link{make_div}()}
}
\concept{div}
\keyword{div}
\keyword{internal}
