% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up2code-package.R
\docType{package}
\name{up2code-package}
\alias{up2code}
\alias{up2code-package}
\title{up2code: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zkamvar/up2code}
  \item Report bugs at \url{https://github.com/zkamvar/up2code/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

}
\keyword{internal}
