% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_headings.R
\name{validate_headings}
\alias{validate_headings}
\title{Validate headings}
\usage{
validate_headings(headings, lesson = NULL, offset = 5L, message = TRUE)
}
\arguments{
\item{headings}{an object of xml_nodelist.}

\item{message}{if \code{TRUE} (default), a message will be issued for each error,
otherwise, they will be silent.}
}
\value{
a logical vector of length five summarizing the results of the
five aspects listed above: \code{TRUE} for valid headings and \code{FALSE} for
invalid headings.
}
\description{
This will validate the following aspects of all headings:
}
\details{
\itemize{
\item first heading starts at level 2 (\code{first_heading_is_second_level})
\item greater than level 1 (\code{all_are_greater_than_first_level})
\item increse sequentially (e.g. no jumps from 2 to 4) (\code{all_are_sequential})
\item have names (\code{all_have_names})
\item unique in their own hierarchy (\code{all_are_unique})
}
}
\note{
This is an internal function implemented for the \link{Episode} and \link{Lesson}
classes.
}
\keyword{internal}
