% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_links.R
\name{validate_links}
\alias{validate_links}
\title{Validate Links in a markdown document}
\usage{
validate_links(yrn, verbose = TRUE)
}
\arguments{
\item{yrn}{a \link[tinkr:yarn]{tinkr::yarn} or \link{Episode} object.}

\item{verbose}{if TRUE (default), messages will be printed as the validator works.}
}
\value{
a data frame containing the links, locations, and indicators if they
passed tests
}
\description{
This function will validate that links do not throw an error in markdown
documents. This will include links to images and will respect robots.txt for
websites.
}
\details{
\subsection{External links}{

These links must start with a valid and secure protocol. This means that we
will enforce HTTPS over HTTP. Any link with HTTP will be flagged. Most
importantly, these links must not return an error code > 399.
}

\subsection{Cross-lesson links}{

These links will have no protocol, but should resolve to the HTML version of
a page and have the correct capitalisation.
}

\subsection{Anchors (aka fragments)}{

Anchors are located at the end of URLs that start with a \verb{#} sign. These are
used to indicate a section of the documenation.
}
}
