% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_md.R
\name{xml_to_md}
\alias{xml_to_md}
\title{Convert xml to markdown}
\usage{
xml_to_md(body, stylesheet = "xml2md_pegboard.xsl", newlines = FALSE)
}
\arguments{
\item{body}{an xml document}

\item{stylesheet}{the name of a stylesheet passed to \code{get_stylesheet}}

\item{newlines}{a logical indicating that newlines (aka softbreaks) should
be inserted between elements (defaults to \code{FALSE}, meaning that no
separator will be added between elements).}
}
\value{
a character vector of length 1
}
\description{
Convert xml to markdown
}
\examples{
cha <- pegboard:::make_div("challenge")
sol <- pegboard:::make_div("solution")
xml2::xml_add_child(cha, xml2::xml_child(sol, 1), .where =  1)
xml2::xml_add_child(cha, xml2::xml_child(sol, 2), .where = 2)
cat(pegboard:::xml_to_md(cha))
}
\keyword{internal}
